+++ src/Pickles/Pickles.TestFrameworks/MsTest/MsTestElementExtensions.cs
@@ -27,7 +27,7 @@ using PicklesDoc.Pickles.ObjectModel;
 
 namespace PicklesDoc.Pickles.TestFrameworks.MsTest
 {
-    internal static class MsTestElementExtensions
+    public static class MsTestElementExtensions
     {
         private const string Failed = "failed";
 
@@ -98,13 +98,13 @@ namespace PicklesDoc.Pickles.TestFrameworks.MsTest
             return document.Root.Descendants(Ns + "UnitTest").Where(s => s.Element(Ns + "Description") != null);
         }
 
-        internal static Guid ExecutionIdElement(this XElement scenario)
+        public static Guid ExecutionIdElement(this XElement scenario)
         {
             //// <UnitTest>
             ////   <Execution id="   the execution id guid   " />
             //// </UnitTest>
 
-            var xElement = scenario.Element(Ns + "Execution");
+            var xElement = scenario?.Element(Ns + "Execution");
 
             return xElement != null ? new Guid(xElement.Attribute("id").Value) : Guid.Empty;
         }